test: consolidate model inheritance coverage
