+++ internal/sessionworker/process_test.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/telos-org/telos/internal/sessionapi"
@@ -156,5 +157,27 @@ func TestStartEpochDoesNotAppendAfterSessionStops(t *testing.T) {
 	}
 	if len(updated.Epochs) != 1 {
 		t.Fatalf("epochs: got %d want 1", len(updated.Epochs))
+	}
+}
+
+func TestWorkerExportsSelectedModelForNestedCLI(t *testing.T) {
+	const model = "anthropic/claude-sonnet-4-6"
+	t.Setenv("TELOS_MODEL", "openai-codex/gpt-5.5")
+	sessionDir := t.TempDir()
+	manifest := &sessionapi.Manifest{
+		SessionKind: sessionapi.KindController,
+		Runtime:     sessionapi.RuntimeCloud,
+	}
+	manifest.Config.Model = model
+	if err := sessionapi.WriteManifest(manifestPath(sessionDir), manifest); err != nil {
+		t.Fatal(err)
+	}
+	environment := map[string]string{}
+	for _, entry := range Env(sessionDir, StartOptions{Runtime: sessionapi.RuntimeCloud}) {
+		key, value, _ := strings.Cut(entry, "=")
+		environment[key] = value
+	}
+	if got := environment["TELOS_MODEL"]; got != model {
+		t.Fatalf("nested CLI model = %q, want %q", got, model)
 	}
 }
