+++ internal/telosd/BUILD.bazel
@@ -34,7 +34,6 @@ go_test(
     srcs = [
         "config_test.go",
         "controller_reconciler_test.go",
-        "inference_defaults_test.go",
         "package_materializer_test.go",
         "platform_skills_test.go",
         "runtime_identity_test.go",
